@@ -1,19 +1,27 @@
 import org.jetbrains.kotlin.gradle.tasks.KotlinCompile
 
+val jvmTarget = "17"
+
 val junitJupiterVersion = "5.8.2"
-val testcontainersVersion = "1.16.3"
+val testcontainersVersion = "1.17.1"
+val rapidsAndRiversVersion = "2022.04.21-09.34.08966130226f"
+val hikariCPVersion = "5.0.1"
+val vaultJdbcVersion = "1.3.9"
+val flywayCoreVersion = "8.5.5"
+val kotliqueryVersion = "1.7.0"
+
 val mainClass = "no.nav.helse.spare.AppKt"
 
 plugins {
-    kotlin("jvm") version "1.6.10"
+    kotlin("jvm") version "1.6.21"
 }
 
 dependencies {
-    implementation("com.github.navikt:rapids-and-rivers:2022.04.05-09.40.11a466d7ac70")
-    implementation("com.zaxxer:HikariCP:5.0.1")
-    implementation("no.nav:vault-jdbc:1.3.9")
-    implementation("org.flywaydb:flyway-core:8.5.5")
-    implementation("com.github.seratch:kotliquery:1.7.0")
+    implementation("com.github.navikt:rapids-and-rivers:$rapidsAndRiversVersion")
+    implementation("com.zaxxer:HikariCP:$hikariCPVersion")
+    implementation("no.nav:vault-jdbc:$vaultJdbcVersion")
+    implementation("org.flywaydb:flyway-core:$flywayCoreVersion")
+    implementation("com.github.seratch:kotliquery:$kotliqueryVersion")
 
     testImplementation("org.junit.jupiter:junit-jupiter-api:$junitJupiterVersion")
     testImplementation("org.junit.jupiter:junit-jupiter-params:$junitJupiterVersion")
@@ -32,11 +40,11 @@ repositories {
 tasks {
 
     named<KotlinCompile>("compileKotlin") {
-        kotlinOptions.jvmTarget = "17"
+        kotlinOptions.jvmTarget = jvmTarget
     }
 
     named<KotlinCompile>("compileTestKotlin") {
-        kotlinOptions.jvmTarget = "17"
+        kotlinOptions.jvmTarget = jvmTarget
     }
 
     withType<Test> {
@@ -66,6 +74,6 @@ tasks {
     }
 
     withType<Wrapper> {
-        gradleVersion = "7.4.1"
+        gradleVersion = "7.4.2"
     }
 }
